Created additional zone to fix the issue
